# Conflicts:
#	LifeLike.Api/Api/Models/Config/ConfigRequestModel.cs
#	LifeLike.Api/Commons/LifeLike.Common/Models/BlobItem.cs
#	LifeLike.Api/LifeLike.CloudService/BlobStorage/BlobStorage.cs
#	LifeLike.Api/LifeLike.CloudService/CosmosDB/DocumentDBRepository.cs
#	LifeLike.Api/LifeLike.CloudService/CosmosDB/UnitOfWork.cs
#	LifeLike.Api/LifeLike.CloudService/LifeLike.CloudService.csproj
#	LifeLike.Api/LifeLike.CloudService/TableStorage/TableStorage.cs
#	LifeLike.Api/LifeLike.Data/LifeLike.Data.csproj
#	LifeLike.Api/LifeLike.Data/Repository.cs
#	LifeLike.Api/LifeLike.Data/UnitOfWork.cs
#	LifeLike.Api/LifeLike.Services/CloudServices/LogService.cs
#	LifeLike.Api/LifeLike.Services/LifeLike.Services.csproj
#	LifeLike.Api/LifeLike.Services/Model/Page.cs
#	LifeLike.Api/LifeLike.Services/Model/Video.cs
#	LifeLike.Api/LifeLike.Services/Profiles/DomainProfile.cs
#	LifeLike.Api/LifeLike.Services/Services/BaseService.cs
#	LifeLike.Api/LifeLike.Services/Services/ConfigService.cs
#	LifeLike.Api/LifeLike.Services/Services/LocalBlobStorage.cs
#	LifeLike.Api/LifeLike.Services/Services/PhotoService.cs
#	LifeLike.Api/LifeLike.Services/Services/VideoService.cs
#	LifeLike.Api/LifeLike.Services/Structures/IPhotoService.cs
#	LifeLike.Api/LifeLike.Shared/IRepository.cs
#	LifeLike.Api/LifeLike.Shared/LifeLike.Shared.csproj
#	LifeLike.Api/LifeLike.Shared/Services/IBlobStorage.cs
#	LifeLike.Api/LifeLike.Shared/Services/ITableStorage.cs
#	LifeLike.Api/LifeLike.Web/Controllers/ConfigController.cs
#	LifeLike.Api/LifeLike.Web/Controllers/LogController.cs
#	LifeLike.Api/LifeLike.Web/Controllers/PageController.cs
#	LifeLike.Api/LifeLike.Web/Controllers/PhotoController.cs
#	LifeLike.Api/LifeLike.Web/Controllers/VideoController.cs
#	LifeLike.Api/LifeLike.Web/LifeLike.Web.csproj
#	LifeLike.Api/LifeLike.Web/LifeLike.Web.csproj.user
#	LifeLike.Api/LifeLike.Web/Program.cs
#	LifeLike.Api/LifeLike.Web/Services/Swagger/Extensions.cs
#	LifeLike.Api/LifeLike.Web/Startup.cs
#	LifeLike.Api/LifeLike.Web/Utils/ImageHelper.cs
#	LifeLike.Api/LifeLike.Web/app.settings.json

+++ LifeLike.Frontend/src/app/modules/admin/pages/photos/photos.component.ts
@@ -1,59 +1,70 @@
 import { Component, OnInit } from '@angular/core';
+import { MatDialog, MatDialogConfig } from '@angular/material';
 import { map } from 'rxjs/operators';
-import { AdminRestService } from '../../services/admin-rest.service';
 import Photo from '../../../../modules/photo/models/Photo';
-import { PhotoCreateComponent } from '../photo-create/photo-create.component';
-import { MatDialogConfig, MatDialog } from '@angular/material';
+import { ConfirmDeleteComponent } from '../../dialogs/confirm-delete/confirm-delete.component';
+import { PhotoCreateComponent } from '../../dialogs/photo-create/photo-create.component';
+import { PhotoEditComponent } from '../../dialogs/photo-edit/photo-edit.component';
+import { AdminRestService } from '../../services/admin-rest.service';
 
 @Component({
   selector: 'app-admin-photos',
   templateUrl: './photos.component.html',
-  styleUrls: ['./photos.component.scss']
+  styleUrls: ['./photos.component.scss'],
 })
 export class PhotosComponent implements OnInit {
   Photos: Photo[];
   SelectedPhoto: Photo;
-  displayedColumns: string[] = ['Id', 'Title', 'Url', 'Created','Actions'];
+  displayedColumns: string[] = ['Id', 'Title', 'Url', 'Created', 'Actions'];
   IsLoading: boolean;
   IsEditMode: boolean;
   IsCreateMode: boolean;
   error: any;
-  constructor(private restService: AdminRestService, private dialog: MatDialog) { }
+  constructor(private readonly restService: AdminRestService, private readonly dialog: MatDialog) { }
 
   Remove(photo: Photo): void {
-    console.log('Remove');
-    console.log(photo);
-    this.IsLoading= true;
-    this.restService.deletePhoto(photo.Id)
-      .subscribe(
-        data => {
-          this.IsLoading = false;
-          this.GetPhotos();
-      },
-      error => {
-          this.error = error;
-          this.IsLoading = false;
-      });
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.data = photo.Title;
+    dialogConfig.autoFocus = true;
+    const dialogRef = this.dialog.open(ConfirmDeleteComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        this.IsLoading = true;
+        this.restService.deletePhoto(photo.Id)
+          .subscribe(
+            () => {
+              this.IsLoading = false;
+              this.GetPhotos();
+          },
+          (error) => {
+              this.error = error;
+              this.IsLoading = false;
+          });
+      }
+    });
   }
   Edit(photo: Photo): void {
-    console.log('Edit');
-    console.log(photo);
+
     const dialogConfig = new MatDialogConfig();
     dialogConfig.data = photo;
     dialogConfig.width = '90%';
-    dialogConfig.height = '90%';
+
     dialogConfig.disableClose = true;
     dialogConfig.autoFocus = true;
-   this.dialog.open(PhotoCreateComponent, dialogConfig );
+    const dialogRef = this.dialog.open(PhotoEditComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((result) => {
+      this.GetPhotos();
+    });
   }
-  Create() : void {
-    console.log('Create');
+  Create(): void {
     const dialogConfig = new MatDialogConfig();
-    dialogConfig.width = '90%';
-    dialogConfig.height = '90%';
     dialogConfig.disableClose = true;
+    dialogConfig.width = '90%';
     dialogConfig.autoFocus = true;
-   this.dialog.open(PhotoCreateComponent, dialogConfig );
+    const dialogRef = this.dialog.open(PhotoCreateComponent, dialogConfig);
+    dialogRef.afterClosed().subscribe((result) => {
+      this.GetPhotos();
+    });
   }
 
   GetPhotos(): void {
@@ -61,15 +72,14 @@ export class PhotosComponent implements OnInit {
       .pipe(
         map((data: Photo[]) => {
           this.IsLoading = false;
-          console.log(data);
           return data;
         }))
       .subscribe((p: Photo[]) => this.Photos = p);
   }
   ngOnInit() {
     this.GetPhotos();
-    this.IsEditMode=false;
-    this.IsCreateMode=false;
+    this.IsEditMode = false;
+    this.IsCreateMode = false;
   }
 
 }
